requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0
pyyaml>=6.0.0
onnxruntime>=1.15.0
dataclasses>=0.6; python_version < '3.7'

